move pps tests to their own file
