+++ jupyter-extension/jupyterlab_pachyderm/tests/__init__.py
@@ -1,4 +1,5 @@
 from pathlib import Path
 
+DEFAULT_PROJECT = "default"
 TEST_NOTEBOOK = Path(__file__).parent.joinpath('data/TestNotebook.ipynb')
 TEST_REQUIREMENTS = TEST_NOTEBOOK.with_name("requirements.txt")
